+++ lib/src/applied_mathematics/statistic/central_tendency.dart
@@ -5,16 +5,13 @@ import '../../discrete_mathematics/linear_algebra/tensor/base/tensor_base.dart';
 import '../../utils/equals.dart';
 import 'exceptions/mean_exception.dart';
 
-/// Class that can computes the mean value of a discrete set of numbers
-class Mean {
-  /// Create instanse of [Mean] with set of positive real numbers
-  Mean(TensorBase numbers) {
-    if (numbers.any((e) => e <= 0)) {
-      throw MeanException('All numbers in set must be greatest than zero!');
-    } else {
-      _set = numbers;
-    }
-  }
+/// **Central tendency** (or measure of central tendency) is a central or 
+/// typical value for a probability distribution. It may also be called 
+/// a center or location of the distribution. Colloquially, measures of 
+/// central tendency are often called averages
+class CentralTendency {
+  /// Create instanse of [CentralTendency] with set of real numbers
+  CentralTendency(this._set);
 
   /// Set of numbers
   TensorBase _set;
@@ -23,7 +20,11 @@ class Mean {
   ///
   /// If [weights] isn't `null` then the weighted arithmetic mean are computed.
   /// If provided, [weights] must have the same shape as the number's set.
-  double arithmetic({TensorBase weights}) {
+  num arithmetic({TensorBase weights}) {
+    if (_set.any((e) => e <= 0)) {
+      throw MeanException('All numbers in set must be greatest than zero!');
+    }
+
     var set = _set.copy();
     final w = weights ?? TensorBase.generate(set.shape, (_) => 1);
 
@@ -41,7 +42,11 @@ class Mean {
   ///
   /// If [weights] isn't `null` then the weighted geometric mean are computed.
   /// If provided, [weights] must have the same shape as the number's set.
-  double geometric({TensorBase weights}) {
+  num geometric({TensorBase weights}) {
+    if (_set.any((e) => e <= 0)) {
+      throw MeanException('All numbers in set must be greatest than zero!');
+    }
+
     final set = _set.copy();
     final w = weights ?? TensorBase.generate(set.shape, (_) => 1);
 
@@ -65,7 +70,11 @@ class Mean {
   ///
   /// If [weights] isn't `null` then the weighted harmonic mean are computed.
   /// If provided, [weights] must have the same shape as the number's set.
-  double harmonic({TensorBase weights}) {
+  num harmonic({TensorBase weights}) {
+    if (_set.any((e) => e <= 0)) {
+      throw MeanException('All numbers in set must be greatest than zero!');
+    }
+
     final set = _set.copy();
     final w = weights ?? TensorBase.generate(set.shape, (_) => 1);
 
@@ -85,13 +94,14 @@ class Mean {
   }
 
   /// Computes quadratic mean of set numbers
-  double quadratic() {
-    final set = _set.map((v) => pow(v, 2));
-    return sqrt(set.reduce((f, s) => f + s) / set.itemsCount);
-  }
+  num quadratic() => generalized(2);
 
   /// Computes generalized (power) mean of set numbers with [degree] (p)
-  double generalized(int degree) {
+  num generalized(int degree) {
+    if (_set.any((e) => e <= 0)) {
+      throw MeanException('All numbers in set must be greatest than zero!');
+    }
+
     if (degree == 0) {
       return geometric();
     }
@@ -99,6 +109,53 @@ class Mean {
     final underRoot = _set.map((n) => pow(n, degree)).reduce((f, s) => f + s) /
         _set.itemsCount;
     return Number(underRoot).rootOf(degree).toDouble();
+  }
+
+  /// Gets minima of the set of numbers - equivalent to [generalized] with 
+  /// degree equal to `-Infinity`
+  num minimum() {
+    final list = _set.toList()
+      ..sort();
+    return list.first;
+  }
+
+  /// Gets maxima of the set of numbers - equivalent to [generalized] with 
+  /// degree equal to `Infinity`
+  num maximum() {
+    final list = _set.toList()
+      ..sort();
+    return list.last;
+  }
+
+  /// Gets the value separating the higher half from the lower half
+  /// of a [_set]
+  num median() {
+    final list = _set.toList();
+    final center =list.length ~/ 2;
+    if (list.length % 2 == 0) {
+      return (list.elementAt(center - 1) + list.elementAt(center)) / 2;
+    } else {
+      return list.elementAt(center);
+    }
+  }
+
+  /// Gets most often appeared values of the [_set]
+  Set<num> mode() {
+    final list = _set.toList();
+    final numbers = list.toSet();
+
+    final result = Set<num>();
+    final maxCount = <num, int>{};
+
+    for (final item in numbers) {
+      maxCount[item] = list.where((e) => e == item).length;
+    }
+
+    var max = maxCount.values.toList()..sort();
+    final maxEntries = maxCount.entries.where((e) => e.value == max.last);
+
+    result.addAll(maxEntries.map((e) => e.key));
+    return result;
   }
 
   @override